+++ fuente/index.js
@@ -5,6 +5,8 @@ const routerCalderas = require('./rutas/calderas.rutas');
 const routerEdificios = require('./rutas/edificios.rutas');
 const routerConstructoras = require('./rutas/constructoras.rutas');
 const routerTecnicos = require('./rutas/tecnicos.rutas');
+const routerClientes = require('./rutas/clientes.rutas');
+
 require('dotenv').config();
 
 const port = process.env.PORT || 8090;
@@ -16,6 +18,7 @@ app.use('/tecnicos', routerTecnicos);
 app.use('/calderas', routerCalderas);
 app.use('/edificios', routerEdificios);
 app.use('/constructoras', routerConstructoras);
+app.use('/clientes', routerClientes);
 
 app.listen(port, () => {
   console.log(`Servidor corriendo en http://localhost:${port}`);
