--- src/datos/calderas.json
@@ -1,32 +1,0 @@
-[
-  {
-    "id": 618,
-    "nombre": "caldera",
-    "descripcion": "A"
-  },
-  {
-    "id": 638,
-    "nombre": "caldera",
-    "descripcion": "D"
-  },
-  {
-    "id": 836,
-    "nombre": "caldera",
-    "descripcion": "C"
-  },
-  {
-    "id": 682,
-    "nombre": "caldera",
-    "descripcion": "B"
-  },
-  {
-    "id": 123,
-    "nombre": "caldera",
-    "descripcion": "A"
-  },
-  {
-    "id": 111,
-    "nombre": "caldera",
-    "descripcion": "B"
-  }
-]
