--- src/datos/constructoras.json
@@ -1,58 +1,0 @@
-[
-  {
-    "id": 908,
-    "nombre": "Fiveclub",
-    "cuit": "883-648-6919",
-    "telefono": "502-695-4997",
-    "edificiosIds": [
-      488,
-      344
-    ]
-  },
-  {
-    "id": 435,
-    "nombre": "Divavu",
-    "cuit": "413-625-0120",
-    "telefono": "505-587-0735",
-    "edificiosIds": [
-      112
-    ]
-  },
-  {
-    "id": 342,
-    "nombre": "Rhynoodle",
-    "cuit": "497-514-1647",
-    "telefono": "576-316-7195",
-    "edificiosIds": [
-      998,
-      768
-    ]
-  },
-  {
-    "id": 717,
-    "nombre": "Jaxnation",
-    "cuit": "292-584-4291",
-    "telefono": "556-444-5522",
-    "edificiosIds": [
-      1,
-      3
-    ]
-  },
-  {
-    "id": 997,
-    "nombre": "Photobean",
-    "cuit": "506-910-5245",
-    "telefono": "152-212-8081",
-    "edificiosIds": [
-      344,
-      4
-    ]
-  },
-  {
-    "id": 9157,
-    "nombre": "giannnn",
-    "cuit": "",
-    "telefono": "",
-    "edificiosIds": []
-  }
-]
