--- src/datos/edificios.json
@@ -1,66 +1,0 @@
-[
-  {
-    "id": 688,
-    "direccion": "",
-    "ciudad": "",
-    "codigo_postal": "",
-    "es_particular": false,
-    "contacto": ""
-  },
-  {
-    "id": 614,
-    "direccion": "40527 Bunker Hill Street",
-    "ciudad": "Maragogipe",
-    "codigo_postal": 344,
-    "es_particular": false,
-    "contacto": "392-879-7706"
-  },
-  {
-    "id": 704,
-    "direccion": "9248 Grover Junction",
-    "ciudad": "Riolândia",
-    "codigo_postal": 650,
-    "es_particular": true,
-    "contacto": "998-985-8262"
-  },
-  {
-    "id": 703,
-    "direccion": "227 Maple Trail",
-    "ciudad": "Gunjur",
-    "codigo_postal": 637,
-    "es_particular": true,
-    "contacto": "183-136-2856"
-  },
-  {
-    "id": 222,
-    "direccion": "9 Old Gate Alley",
-    "ciudad": "Banjarbaru",
-    "codigo_postal": 641,
-    "es_particular": true,
-    "contacto": "221-771-5755"
-  },
-  {
-    "id": 887,
-    "direccion": "53 Eagan Drive",
-    "ciudad": "Maiorca",
-    "codigo_postal": 927,
-    "es_particular": false,
-    "contacto": "924-903-4163"
-  },
-  {
-    "id": 757,
-    "direccion": "3 Pond Pass",
-    "ciudad": "Serra de Água",
-    "codigo_postal": 46,
-    "es_particular": false,
-    "contacto": "635-234-8369"
-  },
-  {
-    "id": 92,
-    "direccion": "91 Algoma Place",
-    "ciudad": "Clermont-Ferrand",
-    "codigo_postal": 673,
-    "es_particular": true,
-    "contacto": "385-897-3114"
-  }
-]
